use Lombok for all the payload types
This is a no-op refactor, which I performed w/ IntelliJ's magic. In
particular, I didn't go to @Data, just @Getter/@Setter, so that equals
and hashcode wouldn't show up and potentially change equality semantics.
Some payload types are @Data already, and switching the rest probably
makes sense, but it wouldn't be a no-op.