correct transactional GraphQL ExecutionStrategy
Using @Transactional places the transaction boundary at the wrong
place, so "marked as rollback-only" is the error message you get back
in the client for _any_ exception raised w/in the transaction. The
logs give you the right message, but not via the API itself.

Instead, manually manage the transaction (via TransactionTemplate) so
the global rollback state can be propagated to the status object and
thus make the rollback explicit. This feels dirty...