add modCount to plan item and bucket
The iCalendar SEQUENCE property is a monotonically increasing 32-bit int
value. We use timestamps for versioning, so can't use that directly. And
need a compound value, as a item's event is based on both the item
itself and its associted bucket. Tracking a modCount and adding them for
the SEQUENCE does what we need.