reimplement upload storage w/out Spring or IOException
Instead of using Spring's MultipartFile directly, wrap it with an Upload
type of our own, and use that for all the storage. Currently, that's
recipes and textract. This is to facilitate supporting the same actions
from GraphQL, which speaks Servlet API (no Spring). Along the way, get
rid of IOException, since we always wrap and rethrow anyway.

For similar reasons, move the recipe photo upload handling out of the
controller and down into the service layer.