explicitly close the input stream after uploading to S3
The S3 client ends up closing the stream, but doesn't guarantee that it
will. In fact, it takes care to put the original stream back (after
replacing it with its digest-y one) for the caller to close as needed.
Be safe and do it explicitly; it's just a no-op if it happens twice.